# swagger.json
{
	"swagger": "2.0",
	"info": {
		"title": {
			"Fn::Sub": "Your Shared Secret API (${Stage})"
		},
		"version": "1.0",
		"description": "API for managing secrets and caretakers for the Your Shared Secret service."
	},
	"host": "api.yoursharedsecret.com",
	"basePath": "/v1",
	"schemes": [
		"https"
	],
	"paths": {
...
		"/caretakers/{secret-id}/{caretaker-id}/send": {
			"post": {
				"operationId": "SendMessage",
				"summary": "Send a message to a caretaker.",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"description": "Unique secret identifier.",
						"name": "secret-id",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "caretaker-id",
						"description": "Unique caretaker identifier.",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "x-yoursharedsecret-ownership",
						"description": "Header proving access to either the private key of the secret in the case of sending an INVITE. If sending an UNLOCK request the private key of the current unlocking caretaker must be used to sign the request. See separate section for details.",
						"in": "header",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "SendRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/SendRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "200 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					},
					"400": {
						"description": "400 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					},
					"404": {
						"description": "404 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					}
				},
		        "x-amazon-apigateway-request-validator": "validator",
				"x-amazon-apigateway-integration": {
					"uri": {
						"Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SecretService.Arn}/invocations"
					},
					"passthroughBehavior": "when_no_match",
					"httpMethod": "POST",
					"type": "aws_proxy"
				}
			}
		},
...
		"/secrets/{secret-id}/{caretaker-id}/unlock": {
			"post": {
				"operationId": "UnlockSecret",
				"summary": "Request to unlock a secret.",
				"consumes": [
					"application/json"
				],
				"produces": [
					"application/json"
				],
				"parameters": [
					{
						"name": "secret-id",
						"description": "Unique secret identifier.",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "caretaker-id",
						"description": "Unique caretaker identifier.",
						"in": "path",
						"required": true,
						"type": "string"
					},
					{
						"name": "x-yoursharedsecret-ownership",
						"description": "Header proving access to the private key of the caretaker requesting the unlock. See separate section for details.",
						"in": "header",
						"required": true,
						"type": "string"
					},
					{
						"in": "body",
						"name": "UnlockSecretRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/UnlockSecretRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "200 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					},
					"404": {
						"description": "404 response",
						"schema": {
							"$ref": "#/definitions/MessageResponse"
						}
					}
				},
		        "x-amazon-apigateway-request-validator": "validator",
				"x-amazon-apigateway-integration": {
					"uri": {
						"Fn::Sub": "arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${SecretService.Arn}/invocations"
					},
					"passthroughBehavior": "when_no_match",
					"httpMethod": "POST",
					"type": "aws_proxy"
				}
			}
		},
...
	},
	"definitions": {
		"MessageResponse": {
			"type": "object",
			"title": "Free form message response.",
			"properties": {
				"message": {
					"description": "Free form message.",
					"type": "string"
				}
			}
		},
...
		"SendRequest": {
			"type": "object",
			"title": "Send message request body payload.",
			"properties": {
				"sendType": {
					"description": "What kind of message to send.\n* INVITE - Initial invite email.\n* UNLOCK - Unlock request email.\n* SHARE - Share notification email.",
					"type": "string",
					"enum": [
						"INVITE",
						"UNLOCK",
						"SHARE"
					]
				},
				"address": {
					"description": "The unencrypted address to send to send to. This must match one of the addresses of the recipient (Checked through the SHA512 hash of the address).",
					"type": "string"
				},
				"title": {
					"description": "Custom title to include.",
					"type": "string"
				},
				"message": {
					"description": "Custom message to include. This message does not allow any formatting.",
					"type": "string"
				}
			},
			"required": [
				"sendType",
				"title",
				"address",
				"message"
			]
		},
...
		"UnlockSecretRequest": {
			"type": "object",
			"title": "Unlock secret request body.",
			"properties": {
				"addressKeyDigests": {
					"type": "object",
					"description": "A map with the key being the caretaker ID of every caretaker of the secret that has a part of the key. The value is a SHA256 digest of the address needed for decrypting the data payload. The value is encoded using Base64 URL safe encoding.",
					"additionalProperties": {
						"type": "string"
					}
				}
			},
			"required": [
				"addressKeyDigests"
			]
		},
...
	},
...
}